chore: move helper functions to a separate file
